main
This is the first iteration of main focusing to testing initial functionality of our equipment by helping us find out which readings can be mapped onto which switched and also includes the beginning of the switch function that we can iterate on later. Also includes some tests for the lcd display.